+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/environment/database/entity/EnvironmentEntity.java
@@ -41,5 +41,5 @@ public class EnvironmentEntity extends BaseEntity {
   public void setEnvironmentSpec(String environmentSpec) {
     this.environmentSpec = environmentSpec;
   }
-  
+
 }
