+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/experiment/database/entity/ExperimentEntity.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.submarine.server.experiment.database;
+package org.apache.submarine.server.experiment.database.entity;
 
 import org.apache.submarine.server.database.entity.BaseEntity;
