+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/notebook/database/mappers/NotebookMapper.java
@@ -17,7 +17,9 @@
  * under the License.
  */
 
-package org.apache.submarine.server.notebook.database;
+package org.apache.submarine.server.notebook.database.mappers;
+
+import org.apache.submarine.server.notebook.database.entity.NotebookEntity;
 
 import java.util.List;
