+++ submarine-server/server-core/src/test/java/org/apache/submarine/server/experiment/ExperimentManagerTest.java
@@ -28,8 +28,8 @@ import org.apache.submarine.server.api.Submitter;
 import org.apache.submarine.server.api.experiment.Experiment;
 import org.apache.submarine.server.api.experiment.ExperimentId;
 import org.apache.submarine.server.api.spec.ExperimentSpec;
-import org.apache.submarine.server.experiment.database.ExperimentEntity;
-import org.apache.submarine.server.experiment.database.ExperimentService;
+import org.apache.submarine.server.experiment.database.entity.ExperimentEntity;
+import org.apache.submarine.server.experiment.database.service.ExperimentService;
 import org.junit.Before;
 import org.junit.Test;
 import org.slf4j.Logger;
